fix: separate fixtures from mocks
