+++ packages/profile-sync-controller/src/sdk/__fixtures__/userstorage.ts
@@ -1,37 +1,15 @@
 import nock from 'nock';
 
-import encryption, { createSHA256Hash } from '../../shared/encryption';
-import { Env } from '../../shared/env';
-import { USER_STORAGE_FEATURE_NAMES } from '../../shared/storage-schema';
-import { STORAGE_URL } from '../user-storage';
+import {
+  MOCK_STORAGE_RESPONSE,
+  MOCK_STORAGE_URL,
+  MOCK_STORAGE_URL_ALL_FEATURE_ENTRIES,
+} from '../mocks/userstorage';
 
 type MockReply = {
   status: nock.StatusCode;
   body?: nock.Body;
 };
-
-// Example mock notifications storage entry (wildcard)
-const MOCK_STORAGE_URL = STORAGE_URL(
-  Env.PRD,
-  `${USER_STORAGE_FEATURE_NAMES.notifications}.notification_settings`,
-);
-const MOCK_STORAGE_URL_ALL_FEATURE_ENTRIES = STORAGE_URL(
-  Env.PRD,
-  USER_STORAGE_FEATURE_NAMES.notifications,
-);
-
-export const MOCK_STORAGE_KEY = createSHA256Hash('mockStorageKey');
-export const MOCK_NOTIFICATIONS_DATA = '{ is_compact: false }';
-export const MOCK_NOTIFICATIONS_DATA_ENCRYPTED = async (data?: string) =>
-  await encryption.encryptString(
-    data ?? MOCK_NOTIFICATIONS_DATA,
-    MOCK_STORAGE_KEY,
-  );
-
-export const MOCK_STORAGE_RESPONSE = async (data?: string) => ({
-  HashedKey: '8485d2c14c333ebca415140a276adaf546619b0efc204586b73a5d400a18a5e2',
-  Data: await MOCK_NOTIFICATIONS_DATA_ENCRYPTED(data),
-});
 
 export const handleMockUserStorageGet = async (mockReply?: MockReply) => {
   const reply = mockReply ?? {
